requirements.txt
pydantic~=2.11.9
numpy~=2.3.3
openai~=1.108.2
faiss-cpu~=1.12.0
